EVM-888: Move TestEnv out.
